requirements.txt
flask
requests
python-dotenv
openai>=1.2.0
gunicorn
